Bump up used Mir version
This is not yet the next release of Mir, just an intermediate version
for development purposes that is close to what the next release will be.
It is expected that this version will soon be updated
to a newer Mir release.

Using this version of mir, Lotus DOES NOT RUN yet,
as the way of announcing new epochs slightly changed.
This should be fixed in another commit.