Adds runtime loading of consensus algorithm via the eudico binary (#52)
This PR allows to select the consensus algorithm at runtime instead of build time. It removes the build.Consensus constant from the code and introduces independent command-line commands for each consensus algorithm. In doing so, it introduces the eudico binary. I purposely left out of this PR replacing the mir-validator binary with the eudico binary. Since this PR is already very large, I'll leave that refactor for a follow-up PR.